% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc.R
\name{process_bc}
\alias{process_bc}
\title{Functions to process datasets: BC}
\usage{
process_bc(uuid, val, fmt, ds, prov, hr, date_current, testing_type)
}
\arguments{
\item{uuid}{The UUID of the dataset from datasets.json.}

\item{val}{The desired value to be extracted from the dataset. Valid
inputs, include "cases", "mortality", "testing", etc. See details.}

\item{fmt}{The format of the output data.}

\item{ds}{The downloaded dataset in the format returned by `dl_dataset`.}
}
\value{
A standardized version of the dataset for the desired value.
}
\description{
Process a dataset listed in datasets.json of Covid19CanadaArchive (https://github.com/ccodwg/Covid19CanadaArchive/blob/master/datasets.json)
into a standardized format for a specific value provided by that dataset
(e.g., cases, deaths, hospitalizations, etc.). The dataset must be provided
via `ds` in the format returned by `dl_dataset` from `Covid19CanadaData`.
}
\details{
The currently supported values are the following:
\itemize{
 \item cases
 \item mortality
 \item recovered
 \item active
 \item testing
 \item vaccine_distribution
 \item vaccine_administration
 \item vaccine_first_doses
 \item vaccine_completion
 \item vaccine_additional_doses
 \item hospitalizations
 \item icu
}
}
\keyword{internal}
