% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk.R
\name{process_sk}
\alias{process_sk}
\title{Functions to process datasets: SK}
\usage{
process_sk(uuid, val, fmt, ds, prov, hr, date_current, testing_type)
}
\arguments{
\item{uuid}{The UUID of the dataset from datasets.json.}

\item{val}{The desired value to be extracted from the dataset. Valid
inputs, include "cases", "mortality", "testing", etc. See details.}

\item{fmt}{The format of the output data.}

\item{ds}{The downloaded dataset in the format returned by `dl_dataset`.}
}
\value{
A standardized version of the dataset for the desired value.
}
\description{
Process a dataset listed in datasets.json of Covid19CanadaArchive (https://github.com/ccodwg/Covid19CanadaArchive/blob/master/datasets.json)
into a standardized format for a specific value provided by that dataset
(e.g., cases, deaths, hospitalizations, etc.). The dataset must be provided
via `ds` in the format returned by `dl_dataset` from `Covid19CanadaData`.
}
\details{
The currently supported values are the following:
\itemize{
 \item cases
 \item mortality
 \item recovered
 \item active
 \item testing
 \item vaccine_distribution
 \item vaccine_coverage_dose_1
 \item vaccine_coverage_dose_2
 \item vaccine_coverage_dose_3
 \item vaccine_coverage_dose_4
 \item vaccine_coverage_dose_5
 \item vaccine_coverage_last_6_months
 \item vaccine_administration_total_doses
 \item vaccine_administration_dose_1
 \item vaccine_administration_dose_2
 \item vaccine_administration_dose_3
 \item vaccine_administration_dose_4
 \item vaccine_administration_dose_5
 \item vaccine_administration_last_6_months
 \item vaccine_total_doses
 \item vaccine_dose_1
 \item vaccine_dose_2
 \item vaccine_dose_3
 \item vaccine_dose_4
 \item vaccine_additional_doses
 \item hospitalizations
 \item hospitalizations_cum
 \item hosp_admissions
 \item icu
 \item icu_admissions
 \item wastewater_copies_per_ml
}
}
\keyword{internal}
