% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_special.R
\name{process_special}
\alias{process_special}
\alias{process_pho_data_tool}
\title{Special data processing functions}
\usage{
process_pho_data_tool()
}
\description{
Some datasets are such that they cannot be processed via the usual pipeline
(using \code{\link[Covid19CanadaDataProcess]{process_dataset}}).
}
