% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{webdriver}
\alias{webdriver}
\alias{webdriver_wait_for_element}
\title{Helper functions for webdriver navigation}
\usage{
webdriver_wait_for_element(webdriver, By, value, timeout)
}
\arguments{
\item{webdriver}{RSelenium server object.}

\item{By}{Type of selector to use.}

\item{value}{Value for selector.}

\item{timeout}{Timeout in seconds.}
}
\description{
Helper functions for webdriver navigation

webdriver_wait_for_element: Wait until element is visible
}
