% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_dataset_helpers}
\alias{process_dataset_helpers}
\alias{helper_cum_current}
\alias{helper_ts}
\alias{helper_ts_can}
\alias{phac_prov}
\title{Helper functions for process_dataset}
\usage{
helper_cum_current(
  .data,
  loc = c("prov", "hr", "subhr"),
  val,
  prov,
  date_current,
  hr = NULL
)

helper_ts(.data, loc = c("prov", "hr"), val, prov, convert_to_cum = FALSE)

helper_ts_can(
  d,
  val,
  convert_to_cum = FALSE,
  val_numeric = FALSE,
  keep_update = FALSE
)

phac_prov(p, mode = c("to_phac", "from_phac"))
}
\arguments{
\item{.data}{The dataset to be processed.}

\item{loc}{One of "prov" or "hr", depending on the spatial resolution.}

\item{val}{The value.}

\item{prov}{The province.}

\item{date_current}{The date provided to cum_current (usually the current date).}

\item{hr}{The health region, if providing data for one specific health region only. Used for ON PHU data only.}

\item{convert_to_cum}{Convert values to cumulative values? Default: FALSE.}

\item{d}{The dataset to be processed.}

\item{val_numeric}{Is the value a float (versus an integer)? Default: FALSE.}

\item{keep_update}{If there is an `update` column, should it be kept? Default: FALSE.}

\item{p}{The column to be processed.}

\item{mode}{One of "to_phac" or "from_phac". Convert between abbreviated
province names (e.g., ON) or the full province names used by PHAC
(e.g., Ontario).}
}
\description{
Helper functions for \code{\link{process_dataset}}
}
