% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_dataset_helpers}
\alias{process_dataset_helpers}
\alias{helper_cum_current}
\alias{helper_ts}
\title{Helper functions for process_dataset}
\usage{
helper_cum_current(.data, loc = c("prov", "hr"), val, prov, date_current)

helper_ts(.data, loc = c("prov", "hr"), val, prov, convert_to_cum = FALSE)
}
\arguments{
\item{.data}{The dataset to be processed.}

\item{loc}{One of "prov" or "hr", depending on the spatial resolution.}

\item{val}{The value.}

\item{prov}{The province.}

\item{date_current}{The date provided to cum_current (usually the current date).}

\item{convert_to_cum}{Convert values to cumulative values? Default: FALSE.}
}
\description{
Helper functions for \code{\link{process_dataset}}
}
